remove .cache

+++ 信息科学与工程学院/电子系统设计实践/src/utils.h
@@ -5,7 +5,7 @@
 
 void empty_loop(unsigned int times);
 void delay_ms(unsigned int ms);
-void delay_us(unsigned int us);
+void delay_us(unsigned long us);
 void overflow_add1(unsigned char *num, unsigned char limit);
 void overflow_sub1(unsigned char *num, unsigned char limit);
